requirements.txt
pandas
redis
hiredis
python-dotenv
